Local llm improvements - load model from app configuration (#124)
