--- src/arduino/app_bricks/large_language_model/README.md
@@ -1,13 +1,0 @@
-# Large Language Model (LLMs) Brick
-
-A Brick for interacting with locally-based Large Language Models (LLMs).
-
-This class wraps LangChain functionality to provide a simplified, unified interface for chatting with models like Qwenm, LLama, Gemma.
-It supports both synchronous 'one-shot' responses and streaming output, with optional conversational memory.
-
-# TODO - backup
-
-https://qaihub-public-assets.s3.us-west-2.amazonaws.com/qai-hub-models/models/qwen2_5_1_5b_instruct/releases/v0.45.0/qwen2_5_1_5b_instruct-genie-w4-qualcomm-qcs8275.zip
-https://qaihub-public-assets.s3.us-west-2.amazonaws.com/qai-hub-models/models/qwen2_5_7b_instruct/releases/v0.45.0/qwen2_5_7b_instruct-genie-w8a16-qualcomm-qcs8275.zip
-https://qaihub-public-assets.s3.us-west-2.amazonaws.com/qai-hub-models/models/whisper_small/releases/v0.45.0/whisper_small-qnn_context_binary-float-qualcomm-qcs8275.zip
-
