--- src/arduino/app_bricks/large_language_model/__init__.py
@@ -1,8 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (C) ARDUINO SRL (http://www.arduino.cc)
-#
-# SPDX-License-Identifier: MPL-2.0
-
-from langchain_core.tools import tool
-from .local_llm import LargeLanguageModel
-
-__all__ = ["LargeLanguageModel", "tool"]
