--- src/arduino/app_bricks/large_language_model/brick_config.yaml
@@ -1,12 +1,0 @@
-id: arduino:large_language_model
-name: Large Language Model (LLM)
-description: "Large Language Model (LLM) Brick enables seamless integration with locally hosted LLMs for advanced AI capabilities in your Arduino projects."
-requires_container: true
-requires_model: true
-model: genie:qwen2.5-7b
-supported_boards: ["unoq", "ventunoq"]
-variables:
-  - name: LOCAL_LLM_API_KEY
-    description: API Key for the local LLM service
-    default_value: api_key
-    secret: true
