+++ src/arduino/app_bricks/llm/brick_compose.unoq.yaml
@@ -18,7 +18,7 @@ services:
       - LLAMA_ARG_BATCH=${LLAMA_ARG_BATCH:-256}
       - LLAMA_ARG_UBATCH=${LLAMA_ARG_UBATCH:-128}
       - LLAMA_ARG_MLOCK=${LLAMA_ARG_MLOCK:-1}
-      - LLAMA_ARG_N_PREDICT=${LLAMA_ARG_N_PREDICT:-2048}
+      - LLAMA_ARG_N_PREDICT=${LLAMA_ARG_N_PREDICT:-256}
       - XDG_CACHE_HOME=/models
       #- LLAMA_ARG_THREADS=${LLAMA_ARG_THREADS:-4}
       #- LLAMA_API_KEY=${LLAMA_API_KEY:-key}
