+++ src/arduino/app_bricks/llm/brick_compose.yaml
