+++ src/arduino/app_bricks/llm/examples/02_simple_streaming_chat.py
@@ -8,7 +8,7 @@
 from arduino.app_bricks.llm import LargeLanguageModel
 from arduino.app_utils import App
 
-llm = LargeLanguageModel()
+llm = LargeLanguageModel(max_tokens=512)
 
 
 def ask_prompt():
