+++ src/arduino/app_bricks/llm/examples/03_external_provider_streaming_chat.py
@@ -8,7 +8,7 @@
 from arduino.app_bricks.llm import LargeLanguageModel
 from arduino.app_utils import App
 
-llm = LargeLanguageModel(model="genie:qwen3:4b")
+llm = LargeLanguageModel(model="qwen3:4b", base_url="http://localhost:11434/v1", max_tokens=512)
 
 
 def ask_prompt():
